Update hyperdeck.js
Adding the option to allow a desired date and time mask to be added to the recording name. 
Recognized masks are: YYYY for year, MM for month, DD for day, HH for hour, MI for minute, '-' as separator "-" and '_' as separator. Any other input is ignored and dumped. 
Default will return the date and time stamp in the existing format.